+++ R/forest-inventory.R
@@ -242,7 +242,8 @@ fd_inventory_spain <- function(province,
                                path_metadata = NULL,
                                quiet         = FALSE) {
   # 0. Handle errors
-  if (!requireNamespace("RODBC", quietly = TRUE)) stop("Package `RODBC` is required to access the inventory data. Please, install it.")
+  if (!requireNamespace("DBI", quietly = TRUE)) stop("Package `DBI` is required to access the inventory data. Please, install it.")
+  if (!requireNamespace("odbc", quietly = TRUE)) stop("Package `odbc` is required to access the inventory data. Please, install it.")
   # 1. Filter province
   ## 1.1. Fix province
   province_fix <- province |>
@@ -318,23 +319,24 @@ fd_inventory_spain <- function(province,
     ## 4.1. File name
     filename <- list.files(dir_unzip, full.names = TRUE, pattern = "\\.accdb$")
     ## 4.2. Connect to DB
-    ## 4.2. Connect to DB
-    conn <- RODBC::odbcConnectAccess2007(filename)
-    ## 4.3. Table names
-    tables_vec <- RODBC::sqlTables(conn) |>
-      dplyr::filter(TABLE_TYPE != "SYSTEM TABLE") |>
-      dplyr::pull(TABLE_NAME)
-    ## 4.4. Read data into a list
+    conn <- odbc::dbConnect(odbc::odbc(),
+                            .connection_string = paste0(
+                              "Driver={Microsoft Access Driver (*.mdb, *.accdb)};",
+                              "Dbq=", filename, ";"
+                            ))
+    ## 4.3. Disconnect on exit
+    on.exit(DBI::dbDisconnect(conn))
+    ## 4.4. Table names
+    tables_vec <- DBI::dbListTables(conn)[!grepl("^MSys", DBI::dbListTables(conn))]
+    ## 4.5. Read data into a list
     data_lst <- purrr::map(
       .x = tables_vec,
-      .f = \(x) RODBC::sqlFetch(conn, x) |>
+      .f = \(x) DBI::dbReadTable(conn, x) |>
         tibble::as_tibble()
     )
-    ## 4.5. Disconnect from DB
-    RODBC::odbcClose(conn)
-    ## 4.5. Rename list
+    ## 4.6. Rename list
     names(data_lst) <- tables_vec
-    ## 4.6. Convert IFN4 PCDatosMap to projected SF (only in field database)
+    ## 4.7. Convert IFN4 PCDatosMap to projected SF (only in field database)
     ## -> IFN3 doesn't have huso column
     if (ifn == 4 & database == "field") {
       data_lst$PCDatosMap_sf <- sf::st_as_sf(